requirements.txt
absl-py==2.1.0
accelerate==1.2.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.8.0
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.1.0
beautifulsoup4==4.12.3
blobfile==3.0.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
comm==0.2.2
contourpy==1.3.0
controlnet-aux==0.0.9
cycler==0.12.1
debugpy==1.8.9
decorator==5.1.1
diffusers==0.29.2
-e git+https://github.com/Darkbblue/generic-diffusion-feature.git@d976d1452f5b6caeccb18afeeadffc47f27869a0#egg=diffusion_feature&subdirectory=feature
distro==1.9.0
docker-pycreds==0.4.0
einops==0.8.0
eval_type_backport==0.2.0
exceptiongroup==1.2.2
executing==2.1.0
filelock==3.16.1
flake8==7.1.1
fonttools==4.55.2
frozenlist==1.5.0
fsspec==2024.10.0
ftfy==6.3.1
gitdb==4.0.11
GitPython==3.1.43
grpcio==1.68.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.26.5
idna==3.10
imageio==2.36.1
importlib_metadata==8.5.0
importlib_resources==6.4.5
iniconfig==2.0.0
invisible-watermark==0.2.0
ipdb==0.13.13
ipykernel==6.29.5
ipython==8.18.1
jedi==0.19.2
Jinja2==3.1.4
jiter==0.8.2
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kaleido==0.2.1
kiwisolver==1.4.7
lazy_loader==0.4
lxml==5.3.0
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.9.3
matplotlib-inline==0.1.7
mccabe==0.7.0
modelscope==1.21.0
mpmath==1.3.0
multidict==6.1.0
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
openai==0.28.0
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.6
plotly==5.24.1
pluggy==1.5.0
prompt_toolkit==3.0.48
propcache==0.3.0
protobuf==5.29.1
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycodestyle==2.12.1
pycryptodomex==3.21.0
pydantic==2.10.3
pydantic_core==2.27.1
pyflakes==3.2.0
Pygments==2.18.0
pyparsing==3.2.0
pytest==8.3.4
pytest-env==1.1.5
python-dateutil==2.9.0.post0
pytz==2024.2
PyWavelets==1.6.0
PyYAML==6.0.2
pyzmq==26.2.0
regex==2024.11.6
requests==2.32.3
safetensors==0.4.5
scikit-image==0.24.0
scikit-learn==1.0.2
scipy==1.13.1
seaborn==0.13.2
sentencepiece==0.2.0
sentry-sdk==2.19.2
setproctitle==1.3.4
six==1.17.0
smmap==5.0.1
sniffio==1.3.1
some-package==0.1
soupsieve==2.6
stack-data==0.6.3
sympy==1.13.1
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
threadpoolctl==3.5.0
tifffile==2024.8.30
timm==0.6.7
tokenizers==0.21.0
tomli==2.2.1
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.47.0
triton==3.1.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wandb==0.19.0
wcwidth==0.2.13
Werkzeug==3.1.3
yarl==1.18.3
zipp==3.21.0
